feat(npcs-page): switched from datagrid to cards
Renamed the whole page to comply with the standard format
